move some tests to ManagedChannelServiceConfigTest
